requirements.txt
streamlit==1.30.0
ydata-profiling==4.7.0
plotly==5.18.0
scikit-learn==1.3.2
pandas>=2.0
seaborn>=0.12
notebook
imbalanced-learn
matplotlib-venn
xgboost
pywaffle